refactor: Controller / ViewController separation
divide controller and view controller so that when it become "only API server" in the future, it will be easier to work
