+++ src/main/kotlin/com/tistory/shanepark/dutypark/member/controller/MemberViewController.kt
@@ -1,12 +1,14 @@
 package com.tistory.shanepark.dutypark.member.controller
 
+import com.tistory.shanepark.dutypark.member.domain.annotation.Login
 import com.tistory.shanepark.dutypark.member.service.MemberService
+import com.tistory.shanepark.dutypark.security.domain.dto.LoginMember
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.GetMapping
 
 @Controller
-class MemberController(
+class MemberViewController(
     val memberService: MemberService
 ) {
 
@@ -14,6 +16,12 @@ class MemberController(
     fun index(model: Model): String {
         model.addAttribute("members", memberService.findAll())
         return "index"
+    }
+
+    @GetMapping("/member")
+    fun memberPage(@Login loginMember: LoginMember, model: Model): String {
+        model.addAttribute("member", loginMember)
+        return "member/member"
     }
 
 }
