+++ src/main/resources/templates/member/member.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<head th:replace="~{layout/header::link}"></head>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Duty Park</title>
+</head>
+<body>
+<head th:replace="~{layout/header::header(${'회원 관리'})}"></head>
+<div class="memberPage">
+<!--  TODO : 회원 관리 페이지 생성-->
+    <a th:href="@{/duty/{name}(name=${member.name})}" th:text="${ '(' + member.departmentName + ') ' + member.name}"></a>
+</div>
+</body>
+</html>
